Stand: Auto fährt Punkte an.
Ausweichverhalten funktioniert, Punkte anfahren funktioniert. Noch zu
ungenau, Odometrieräder nicht genau genug. Ab 8.12.2017 gemeisamer Test
mit Team Positionsbestimmung.

+++ Kommunikation.h
@@ -2,7 +2,7 @@
 #define KOMMUNIKATION_H
 #include <Arduino.h>
 
-// TODO: Die Byte LÃ¤ngen sind alle noch auf mindestens 45 zu setzen
+// TODO: Die Byte Längen sind alle noch auf mindestens 45 zu setzen
 class Kommunikation {
 private:
 
@@ -20,7 +20,7 @@ public:
 	Kommunikation();
 	//bool getSignalUsefull();
 	bool getStopEnemy();
-	bool getPosition(int& xPos, int& yPos);
+	bool getPosition(float& xPos, float& yPos);
 
 	// Zum testen Objekt im Robi.cpp anlegen und dann
 	// diese Funktion einfach im loop laufen lassen
@@ -33,4 +33,3 @@ public:
 };
 
 #endif /* KOMMUNIKATION_H */
-
